requirements.txt
prisma
python-dotenv
requests
brotli
beautifulsoup4
python-slugify
Pillow
boto3
aiolimiter
APScheduler
